iter models and switch over to CE loss

+++ mlreco/utils/gnn/data.py
@@ -64,14 +64,14 @@ def cluster_edge_features(data, clusts, edge_index, cuda=True):
     return e
 
 
-def edge_assignment(edge_index, batches, groups, cuda=True):
+def edge_assignment(edge_index, batches, groups, cuda=True, dtype=torch.float):
     """
     edge assignment as same group/different group
     """
     edge_assn = torch.tensor([np.logical_and(
         batches[edge_index[0,k]] == batches[edge_index[1,k]],
         groups[edge_index[0,k]] == groups[edge_index[1,k]]) for k in range(edge_index.shape[1])], 
-                             dtype=torch.float, requires_grad=False)
+                             dtype=dtype, requires_grad=False)
     if cuda:
         edge_assn = edge_assn.cuda()
     return edge_assn
